// plugins/functions/analytics-svc/src/utils/envConverter/types.ts
type ConfigTypes = {
    HANA: "HANA";
    POSTGRES: "POSTGRES";
    BIGQUERY: "BIGQUERY";
    AUDIT: "AUDIT";
    INT_TEST: "INT_TEST";
};

export type CombinedEnv = (
    | HanaConfig
    | PostgresConfig
    | BigqueryConfig
    | IntTestConfig
)[];

export type HanaValues = {
    credentials: {
        user: string;
        password: string;
        adminUser: string;
        adminPassword: string;
        readUser: string;
        readPassword: string;
        customUser: string;
        customPassword: string;
        writeUser: string;
        writePassword: string;
    };
    host: string;
    port: string;
    code: string;
    databaseName: string;
    validateCertificate: boolean;
    pooling: boolean;
    autoCommit: boolean;
    encrypt: boolean;
    enableAuditPolicies: boolean;
    sslTrustStore: string;
    ca: string;
    useTLS: boolean;
    rejectUnauthorized: boolean;
    probeSchema: string;
    vocabSchema: string;
    configSchema: string;
    cdwSchema: string;
    schema: string;
    dialect: string;
    sslCryptoProvider: string;
    hostnameInCertificate: string;
};
export type HanaConfig = {
    type: ConfigTypes["HANA"];
    tags: string[];
    name: string;
    key: string;
    description: string;
    values: HanaValues;
    analyticsSvcValues?: Partial<HanaValues>;
    dbSvcValues?: Partial<HanaValues>;
    cdwSvcValues?: Partial<HanaValues>;
};

export type PostgresValues = {
    credentials: {
        user: string;
        password: string;
        adminUser: string;
        adminPassword: string;
        readUser: string;
        readPassword: string;
        customUser: string;
        customPassword: string;
        writeUser: string;
        writePassword: string;
    };
    host: string;
    port: string;
    code: string;
    database: string;
    databaseName: string;
    schema: string;
    vocabSchema: string;
    connectionTimeoutMillis: number;
    idleInTransactionSessionTimeout: number;
    queryTimeout: number;
    statementTimeout: number;
    dialect: string;
    max: number;
    idleTimeoutMillis: number;
};

export type PostgresConfig = {
    type: ConfigTypes["POSTGRES"];
    tags: string[];
    name: string;
    key: string;
    description: string;
    values: PostgresValues;
    analyticsSvcValues?: Partial<PostgresValues>;
    dbSvcValues?: Partial<PostgresValues>;
};

export type BigqueryValues = {
    code: string;
    id: string;
    host: string;
    vocab_schemas: string[];
    publications: string[];
    authentication_mode: string;
    database: string;
    databaseName: string;
    dialect: string;
    port: string;
    type: string;
    auth_uri: string;
    client_id: string;
    token_uri: string;
    project_id: string;
    private_key: string;
    client_email: string;
    private_key_id: string;
    universe_domain: string;
    client_x509_cert_url: string;
    auth_provider_x509_cert_url: string;
    credentials: {};
};

export type BigqueryConfig = {
    type: ConfigTypes["BIGQUERY"];
    tags: string[];
    name: string;
    key: string;
    description: string;
    values: BigqueryValues;
    analyticsSvcValues?: Partial<PostgresValues>;
    dbSvcValues?: Partial<PostgresValues>;
};

export type IntTestValues = {
    HANASERVER: string;
    TESTPORT: number;
    HDIUSER: string;
    TESTSYSTEMPW: string;
    TESTSCHEMA: string;
};

export type IntTestConfig = {
    type: ConfigTypes["INT_TEST"];
    tags: string[];
    name: string;
    key: string;
    description: string;
    values: IntTestValues;
    analyticsSvcValues?: Partial<IntTestValues>;
    dbSvcValues?: Partial<IntTestValues>;
};